Refactored some code
For each package of Logic classes i made a separate package in test file, so there is no confusion (Also to test if the quality gate works)
