--- app/src/test/java/com/example/piatinkpartyapp/ExampleUnitTest.java
@@ -1,11 +1,0 @@
-package com.example.piatinkpartyapp;
-
-
-/**
- * Example local unit test, which will execute on the development machine (host).
- *
- * @see <a href="http://d.android.com/tools/testing">Testing documentation</a>
- */
-public class ExampleUnitTest {
-
-}
