+++ app/src/test/java/com/example/piatinkpartyapp/cardsTests/CardTest.java
@@ -1,4 +1,4 @@
-package com.example.piatinkpartyapp;
+package com.example.piatinkpartyapp.cardsTests;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
