+++ app/src/test/java/com/example/piatinkpartyapp/gameLogicTests/WattnTest.java
@@ -1,4 +1,4 @@
-package com.example.piatinkpartyapp;
+package com.example.piatinkpartyapp.gameLogicTests;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
